backend/requirements.txt
absl-py==2.4.0
albucore==0.0.24
albumentations==2.0.8
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
beautifulsoup4==4.14.3
brotli==1.2.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cuda-bindings==12.9.4
cuda-pathfinder==1.3.5
cycler==0.12.1
Cython==3.2.4
ddgs==9.10.0
Deprecated==1.3.1
distro==1.9.0
easydict==1.13
fake-useragent==2.2.0
fastapi==0.129.2
filelock==3.20.0
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2025.12.0
google_search_results==2.4.2
googlesearch-python==1.3.0
groq==0.37.1
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.4.1
hyperframe==6.1.0
idna==3.11
ImageIO==2.37.2
insightface==0.7.3
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.9
langchain==1.2.10
langchain-core==1.2.14
langchain-groq==1.1.2
langgraph==1.0.9
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.8
langsmith==0.7.6
lazy_loader==0.4
limits==5.8.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
onnx==1.20.1
onnxruntime-gpu==1.24.2
opencv-contrib-python==4.13.0.92
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pillow==12.1.1
prettytable==3.17.0
primp==1.1.1
protobuf==7.34.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
PyYAML==6.0.3
regex==2026.2.19
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
safetensors==0.7.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.0
setuptools==70.2.0
shellingham==1.5.4
simsimd==6.5.13
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
socksio==1.0.0
sounddevice==0.5.5
soupsieve==2.8.3
starlette==0.52.1
stringzilla==4.6.0
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tifffile==2026.2.24
timm==1.0.24
tokenizers==0.22.2
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
transformers==5.2.0
triton==3.6.0
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.41.0
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
wrapt==2.1.1
xxhash==3.6.0
zstandard==0.25.0
